requirements.txt
# ---------------- Core Backend ----------------
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1

# ---------------- Scraping ----------------
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
urllib3==2.1.0

# ---------------- Vector Search ----------------
faiss-cpu==1.7.4
numpy==1.26.4

# ---------------- NLP / Embeddings ----------------
sentence-transformers==2.2.2
transformers==4.30.0

# ---------------- PyTorch (CPU Only) ----------------
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# ---------------- LLM (Local) ----------------
gpt4all==2.8.2

# ---------------- Utilities ----------------
tqdm==4.66.4

huggingface_hub==0.25.2
